Document comments
- Added comments and documenttext to many components, types and functions
- Renamed and moved the component for the map pins
