+++ frontend/components/functions/AutosizeTextArea.ts
@@ -1,6 +1,10 @@
 import { useEffect } from "react";
 
-// Updates the height of a <textarea> when the value changes.
+/**
+ * Function for updating the height of textareas when the text-input changes
+ * @param textAreaRef Ref to the textarea to update
+ * @param value Text value related to the textarea
+ */
 const autosizeTextArea = (
   textAreaRef: HTMLTextAreaElement | null,
   value: string
