+++ frontend/components/types/MapMarker.ts
@@ -1,3 +1,6 @@
+/**
+ * For use in marker-related logic, as it has marker-related information
+ */
 export type MapMarker = {
   displayName: string;
   latitude: number;
